Add files via upload
Run the React App:

After installing the dependencies, you can start the development server by running:
bash
Copy code
<code>npm start</code>
This command starts the development server and opens your React app in a new browser window.
Access the React App:

Open your web browser and navigate to http://localhost:3000/ (or another port if specified in the terminal). This is the default address where the React app will be running.
Explore and Modify:

Once the app is running, you can explore the functionality. If it's a development server, it will likely auto-reload when you make changes to the source code.
Stop the Development Server:

To stop the development server, go back to the terminal and press Ctrl + C. Confirm by typing 'Y' and hitting Enter.